Fix chan_modem_i4l (bug #3410)
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@4885 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ include/asterisk/alaw.h
@@ -29,6 +29,6 @@ extern unsigned char __ast_lin2a[8192];
 extern short __ast_alaw[256];
 
 #define AST_LIN2A(a) (__ast_lin2a[((unsigned short)(a)) >> 3])
-#define AST_ALAW(a) (__ast_alaw[(a)])
+#define AST_ALAW(a) (__ast_alaw[(int)(a)])
 
 #endif
